Add files via upload
This is R script containing the R Stan code. It uses the stan library to fit the model based on the stan file ("double_binomial.stan"). We create the data_list consisting of the number of observations, N; the observed data, y; and the number of trials, m. The stan function runs the model that has been specified in the stan file ("double_binomial.stan"). We set the total number of iterations, iter to 110,000 with 1000 iterations that will be discarded as warmup and a thinning interval of 10 to minimize autocorrelation. The number of chains used for the sampling was set to 2.