Add files via upload
The data for our analysis was obtained from the developmental toxicity study of ethylene glycol (EG) on the Swiss CD-1 mice conducted through the National Toxicology
Program (NTP). The data are available at http://ntp.niehs.nih.gov. The mice were randomized to either control or one of the three dose levels (750, 1500 or 3000
mg/kg) of the ethylene glycol. For easy and better computational purposes, each dose is divided by the highest dose to scale it down. Outcomes measured included
fetal weight, sex, dead or alive and various forms of malformations. 
the summary statistics of the data. The summary includes the number of clusters
(litters) in each dose group, the number of fetal malformations in each dose group.
From the table, the observed rates of fetal malformation increases with dose from
0.045 for the control group to 0.487 at the largest dose group